Subversion 1.6.9 public header files (directory subversion\include)
My initial plan was to update svn_client.pas and SvnClient.pas to Subversion 1.7, because the Subversion team planned to create a branch for Subversion 1.7 in January 2010. According to the thread "Setting a 1.7 branch goal" in the Subversion development mailing list (February 2010) there isn't yet a date for the 1.7 branch and thatswhy I think it's better to update svn_client.pas and SvnClient.pas to Subversion 1.6 first.